requirements-dev.txt
pytest==8.3.3
coverage==7.6.4
requests==2.32.3
cryptography==43.0.3
psycopg2-binary==2.9.9
